WIP: updated scifio-devel dependencies
Updated scifio-devel to the 0.1.0 snapshot jar. This should be changed
to a release jar, using an updated pom-scijava groupId and version macro
once the "scifio" name clash is resolved.

NB: eventually the IO component should be migrated completely to SCIFIO.
However, there are a huge number of examples and tests that use
ImgOpener, thus it is not practical for the short term.