WIP Eliminate Unit concept in favor of Calibrated
Before there were marker interfaces that had no methods but union interfaces
from the Typed and Unit hierarchies. Now Unit has become Calibrated and it
is always derived from Typed.

For now the calibrated classes are named as CalibratedSpaceNew to avoid the
type name clash with the existing Imglib CalibratedSpace interface (which is
now deprecated). This name clash still needs to be handled.